set up for nonAnonClass
