--- app/src/main/java/edu/umsl/bekvff/contactslist/SingleFragmentActivity.java
@@ -1,28 +1,0 @@
-package edu.umsl.bekvff.contactslist;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
-
-/**
- * Created by Brian Koehler on 4/6/2016.
- */
-public abstract class SingleFragmentActivity extends FragmentActivity{
-    protected abstract Fragment createFragment();
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_fragment);
-
-        FragmentManager manager = getSupportFragmentManager();
-        Fragment fragment = manager.findFragmentById(R.id.fragment_container);
-        if(fragment == null) {
-            fragment = createFragment();
-            manager.beginTransaction()
-                    .add(R.id.fragment_container, fragment)
-                    .commit();
-        }
-    }
-}
